math refactor
