+++ engine/src/core/dm_math.c
@@ -1,27 +1,17 @@
-#include "dm_math_misc.h"
+#include "dm_math.h"
 #include <math.h>
-
-float dm_rad_to_deg(float radians)
-{
-	return  (float)(radians * (180.0f / DM_MATH_PI));
-}
-
-float dm_deg_to_rad(float degrees)
-{
-	return (float)(degrees * (DM_MATH_PI / 180.0f));
-}
 
 float dm_math_angle_xy(float x, float y)
 {
 	float theta = 0.0f;
-
+    
 	if (x >= 0.0f)
 	{
 		theta = atanf(y / x);
 		if (theta < 0.0f) theta += 2.0f * DM_MATH_PI;
 	}
 	else theta = atanf(y / x) + DM_MATH_PI;
-
+    
 	return theta;
 }
 
@@ -34,6 +24,7 @@ float dm_cos(float angle)
 {
 	return cosf(angle);
 }
+
 
 float dm_tan(float angle)
 {
