--- engine/src/core/math/dm_math.h
@@ -1,14 +1,0 @@
-#ifndef __DM_MATH_H__
-#define __DM_MATH_H__
-
-#include "dm_math_misc.h"
-#include "dm_math_vec2.h"
-#include "dm_math_vec3.h"
-#include "dm_math_vec4.h"
-#include "dm_math_quat.h"
-#include "dm_math_mat2.h"
-#include "dm_math_mat3.h"
-#include "dm_math_mat4.h"
-#include "dm_math_mat_transforms.h"
-
-#endif
